+++ src/Config/DatabaseConfig.php
@@ -24,7 +24,7 @@ final class DatabaseConfig extends InjectableConfig
     public const CONFIG = 'database';
     public const DEFAULT_DATABASE = 'default';
 
-    public function __construct(array $config)
+    public function __construct(array $config = [])
     {
         parent::__construct(\array_merge([
             'default' => self::DEFAULT_DATABASE,