+++ tests/Database/Stub/Driver/MSSQLWrapDriver.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\Database\Tests\Stub\Driver;
+
+use Cycle\Database\Driver\SQLServer\SQLServerDriver;
+
+class MSSQLWrapDriver extends SQLServerDriver
+{
+    use TestDriverTrait;
+}