+++ tests/Database/Stub/UntypedEnum.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\Database\Tests\Stub;
+
+enum UntypedEnum
+{
+    case FOO;
+    case BAR;
+    case BAZ;
+}