+++ tests/Database/Unit/Query/InterpolatorTest.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace Cycle\Database\Tests\Unit\Query;
 
-use PHPUnit\Framework\TestCase;
 use Cycle\Database\Injection\Parameter;
 use Cycle\Database\Query\Interpolator;
+use Cycle\Database\Tests\Stub\FooBarEnum;
+use Cycle\Database\Tests\Stub\IntegerEnum;
+use Cycle\Database\Tests\Stub\UntypedEnum;
+use PHPUnit\Framework\TestCase;
 use stdClass;
 
 class InterpolatorTest extends TestCase
@@ -27,6 +30,45 @@ class InterpolatorTest extends TestCase
 
         $this->assertSame(
             'SELECT * FROM table WHERE name = \'Anton\' AND id IN(1, 2, 3) AND balance > 120',
+            $interpolated
+        );
+    }
+
+    /**
+     * @requires PHP >= 8.1
+     */
+    public function testEnumInterpolation(): void
+    {
+        $query = 'SELECT * FROM table WHERE enums = :str OR enumi = :int';
+
+        $parameters = [
+            ':str' => FooBarEnum::BAR,
+            ':int' => IntegerEnum::ANSWER,
+        ];
+
+        $interpolated = Interpolator::interpolate($query, $parameters);
+
+        $this->assertSame(
+            "SELECT * FROM table WHERE enums = 'bar' OR enumi = 42",
+            $interpolated
+        );
+    }
+
+    /**
+     * @requires PHP >= 8.1
+     */
+    public function testUntypedEnumInterpolation(): void
+    {
+        $query = 'SELECT * FROM table WHERE enum = :enum';
+
+        $parameters = [
+            ':enum' => UntypedEnum::BAR,
+        ];
+
+        $interpolated = Interpolator::interpolate($query, $parameters);
+
+        $this->assertSame(
+            'SELECT * FROM table WHERE enum = [UNRESOLVED]',
             $interpolated
         );
     }